requirements.txt
gradio
pdfplumber
sentence-transformers
python-dotenv
google-generativeai